requirements.txt
Flask>=2.0
python-dotenv>=0.20
twilio>=8.0
openai>=0.27
gunicorn>=20.0
requests>=2.28